Refactor lambda post hook.
This allows us to reuse this framework in other places.

Differential Revision: [D20958327](https://our.internmc.facebook.com/intern/diff/D20958327/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D20958327/)!

[ghstack-poisoned]
